requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.6
google-generativeai==0.3.2
pyttsx3==2.90
SpeechRecognition==3.10.1
Pillow==10.2.0
easyocr==1.7.1
fpdf==1.7.2
sqlalchemy==2.0.25
python-dotenv==1.0.1
pydantic==2.6.1 